Move SynapseGroup to synapses.py
